\name{dispFuncs}
\alias{dispFuncs}
\docType{data}
\title{
Dispersion functions
}
\description{
Dispersion functions. EdgeR ? To complete by Andrea
}
\usage{data(dispFuncs)}
\format{
  The format is:
List of 2
 $ : num [1:2] 0.024 14.896
 $ : num [1:2] 0.00557 1.54247
}
\details{
To complete
}
\source{
To complete if necessary
}
\references{
A. Rau, G. Marot and F. Jaffrezic (2012). Differential meta-analysis of RNA-seq data. (Manuscript in progress)
}
\examples{
data(dispFuncs)
}
\keyword{datasets}
