\name{metaRNASeq-package}
\alias{metaRNASeq-package}
\alias{metaRNASeq}
\docType{package}
\title{
Meta-analysis for RNAseq data.
}
\description{
This package implements two p-value combination techniques (inverse normal and Fisher methods). It also provides a vignette explaining how to perform a meta-analysis from two RANseq experiments.
}
\details{
\tabular{ll}{
Package: \tab metaRNASeq\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-11-29\cr
License: \tab GPL \cr
}
}
\author{
Andrea Rau, Guillemette Marot, Florence Jaffrezic 

Maintainer: Guillemette Marot <guillemette.marot@inria.fr>

}
\references{
A. Rau, G. Marot and F. Jaffrezic (2012). Differential meta-analysis of RNA-seq data. (Manuscript in progress)
}
\keyword{ package }
\seealso{
\code{\link{invnorm}}
\code{\link{fishercomb}}
}
\examples{
###user guide
%vignette("metaRNASeq")
}
